WIP: Refactor tests broken during new data util development
