+++ data/types.js
@@ -15,29 +15,32 @@
  * == BSD2 LICENSE ==
  */
 
-/* eslint-disable no-bitwise, no-mixed-operators */
+/* eslint-disable no-bitwise, no-mixed-operators, max-len */
 import _ from 'lodash';
 
 import { addDuration } from '../src/utils/datetime';
-import { MGDL_UNITS, MS_IN_DAY } from '../src/utils/constants';
+import { MGDL_UNITS, MS_IN_DAY, MMOLL_UNITS, MGDL_PER_MMOLL } from '../src/utils/constants';
 
 const APPEND = '.000Z';
+
+export const generateGUID = () => 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, c => {
+  const r = Math.random() * 16 | 0;
+  const v = c === 'x' ? r : (r & 0x3 | 0x8);
+  return v.toString(16);
+});
 
 class Common {
   constructor(opts = {}) {
     this.deviceId = 'Test Page Data - 123';
     this.source = opts.source || 'testpage';
     this.conversionOffset = 0;
+    this.uploadId = opts.uploadId || 'uploadId123';
 
     this.assignGUID();
   }
 
   assignGUID() {
-    const guid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, c => {
-      const r = Math.random() * 16 | 0;
-      const v = c === 'x' ? r : (r & 0x3 | 0x8);
-      return v.toString(16);
-    });
+    const guid = generateGUID();
 
     this.id = guid;
   }
@@ -96,8 +99,8 @@ export class Basal extends Common {
 
     this.time = this.makeTime();
     this.timezoneOffset = this.makeTimezoneOffset();
-    this.normalTime = this.makeNormalTime();
-    this.normalEnd = addDuration(this.normalTime, this.duration);
+    if (!opts.raw) this.normalTime = this.makeNormalTime();
+    if (!opts.raw) this.normalEnd = addDuration(this.normalTime, this.duration);
   }
 }
 
@@ -121,7 +124,7 @@ export class Bolus extends Common {
 
     this.time = this.makeTime();
     this.timezoneOffset = this.makeTimezoneOffset();
-    this.normalTime = this.makeNormalTime();
+    if (!opts.raw) this.normalTime = this.makeNormalTime();
   }
 }
 
@@ -145,7 +148,14 @@ export class CBG extends Common {
 
     this.time = this.makeTime();
     this.timezoneOffset = this.makeTimezoneOffset();
-    this.normalTime = this.makeNormalTime();
+    if (opts.raw) {
+      if (this.units === MGDL_UNITS) {
+        this.units = MMOLL_UNITS;
+        this.value = this.value / MGDL_PER_MMOLL;
+      }
+    } else {
+      this.normalTime = this.makeNormalTime();
+    }
   }
 }
 
@@ -154,21 +164,27 @@ export class Message extends Common {
     super(opts);
 
     _.defaults(opts, {
-      messageText: 'This is a note.',
-      parentMessage: null,
+      messagetext: 'This is a note.',
+      parentmessage: null,
       time: new Date().toISOString(),
     });
 
     this.type = 'message';
 
     this.time = opts.time;
-    const dt = new Date(this.time);
-    const offsetMinutes = dt.getTimezoneOffset();
-    dt.setUTCMinutes(dt.getUTCMinutes() - offsetMinutes);
-    this.normalTime = dt.toISOString();
 
-    this.messageText = opts.messageText;
-    this.parentMessage = opts.parentMessage;
+    if (!opts.raw) {
+      const dt = new Date(this.time);
+      const offsetMinutes = dt.getTimezoneOffset();
+      dt.setUTCMinutes(dt.getUTCMinutes() - offsetMinutes);
+
+      this.normalTime = dt.toISOString();
+      this.messageText = opts.messagetext;
+      this.parentMessage = opts.parentmessage;
+    } else {
+      this.messagetext = opts.messagetext;
+      this.parentmessage = opts.parentmessage;
+    }
   }
 }
 
@@ -217,7 +233,15 @@ export class Settings extends Common {
 
     this.time = this.makeTime();
     this.timezoneOffset = this.makeTimezoneOffset();
-    this.normalTime = this.makeNormalTime();
+    if (opts.raw) {
+      this.units.bg = MMOLL_UNITS;
+      if (this.insulinSensitivity) this.insulinSensitivity.amount = this.insulinSensitivity.amount / MGDL_PER_MMOLL;
+      if (this.bgTarget.target) this.bgTarget.target = this.bgTarget.target / MGDL_PER_MMOLL;
+      if (this.bgTarget.low) this.bgTarget.low = this.bgTarget.low / MGDL_PER_MMOLL;
+      if (this.bgTarget.high) this.bgTarget.high = this.bgTarget.high / MGDL_PER_MMOLL;
+    } else {
+      this.normalTime = this.makeNormalTime();
+    }
   }
 }
 
@@ -241,7 +265,13 @@ export class SMBG extends Common {
     this.time = this.makeTime();
     this.timezoneOffset = this.makeTimezoneOffset();
     this.displayOffset = opts.displayOffset;
-    this.normalTime = this.makeNormalTime();
+
+    if (opts.raw) {
+      this.units = MMOLL_UNITS;
+      this.value = this.value / MGDL_PER_MMOLL;
+    } else {
+      this.normalTime = this.makeNormalTime();
+    }
   }
 }
 
@@ -268,7 +298,7 @@ export class DeviceEvent extends Common {
     this.time = this.makeTime();
     this.createdTime = this.makeTime();
     this.timezoneOffset = this.makeTimezoneOffset();
-    this.normalTime = this.makeNormalTime();
+    if (!opts.raw) this.normalTime = this.makeNormalTime();
   }
 }
 
@@ -289,7 +319,7 @@ export class Upload extends Common {
 
     this.time = this.makeTime();
     this.timezone = opts.timezone;
-    this.normalTime = this.makeNormalTime();
+    if (!opts.raw) this.normalTime = this.makeNormalTime();
     this.createdTime = this.makeTime();
     this.timezoneOffset = this.makeTimezoneOffset();
   }
@@ -313,6 +343,7 @@ export class Wizard extends Common {
       insulinSensitivity: 50,
       recommended: {},
       value: 5.0,
+      units: MGDL_UNITS,
     });
 
     this.type = 'wizard';
@@ -331,7 +362,19 @@ export class Wizard extends Common {
 
     this.time = this.makeTime();
     this.timezoneOffset = this.makeTimezoneOffset();
-    this.normalTime = this.makeNormalTime();
+    if (opts.raw) {
+      this.bolus = this.bolus.id;
+
+      this.units = MMOLL_UNITS;
+      if (this.insulinSensitivity) this.insulinSensitivity = this.insulinSensitivity / MGDL_PER_MMOLL;
+      if (this.bgInput) this.bgInput = this.bgInput / MGDL_PER_MMOLL;
+      if (this.bgTarget.target) this.bgTarget.target = this.bgTarget.target / MGDL_PER_MMOLL;
+      if (this.bgTarget.range) this.bgTarget.range = this.bgTarget.range / MGDL_PER_MMOLL;
+      if (this.bgTarget.low) this.bgTarget.low = this.bgTarget.low / MGDL_PER_MMOLL;
+      if (this.bgTarget.high) this.bgTarget.high = this.bgTarget.high / MGDL_PER_MMOLL;
+    } else {
+      this.normalTime = this.makeNormalTime();
+    }
   }
 }
 
@@ -348,7 +391,7 @@ export class Food extends Common {
     this.nutrition = opts.nutrition;
 
     this.time = this.makeTime();
-    this.normalTime = this.makeNormalTime();
+    if (!opts.raw) this.normalTime = this.makeNormalTime();
     this.createdTime = this.makeTime();
     this.timezoneOffset = this.makeTimezoneOffset();
   }
